Initial refactor as per shin-'s suggestion
